==> 1-Google Geocode API.ipynb <==
# Building the Geocode table:
# First, generate the API URLs
# Then call them and save the json in a list

import requests
import google_api_key # you need the file in this folder
google_geocode_url = "https://maps.googleapis.com/maps/api/geocode/json"

""" input:  school name: str
    output: json from Google Geocode API for this school."""
def get_school_geocode(school_name):
    p = {'address': school_name,
         'key':   google_api_key.key }
    r = requests.get(url = google_geocode_url, params = p)
    data = r.json()
    # TODO make sure this works!
    return data

def get_lat_lng(school_json, lat_or_lng='lat'):
    try:
        return school_json['results'][0]['geometry']['location'][lat_or_lng]
    except:
        return 0.0  # general catchall

import pickle
# these are the pickled versions of the 
# school, degree, and degree_grant tables.
# we heavily cleaned the degree table in step 0.

#with open('school.pickle', 'rb') as f:
#    school = pickle.load(f)
#with open('degree.pickle', 'rb') as f:
#    degree = pickle.load(f)
#with open('degree_grant.pickle', 'rb') as f:
#    degree_grant = pickle.load(f)

# this line attempts to collect all Google 

#school['geocode_json'] = list(map(get_school_geocode, school['school_name']))

# lat/lng can be found at ['results'][0]['geometry']['location']['lat'] and ['lng']

#school.head()

#school['lat'] = list(map(lambda x: get_lat_lng(x, 'lat'), school['geocode_json']))
#school['lng'] = list(map(lambda x: get_lat_lng(x, 'lng'), school['geocode_json']))

#with open('school_geocode_df.pickle', 'wb') as f:
#    pickle.dump(school, f)
#with open('degree_df.pickle', 'wb') as f:
#    pickle.dump(degree, f)

# some schools did not show up in the search. 
# we note them here.
school_old_name = school[school['geocode_json']=={'results': [], 'status': 'ZERO_RESULTS'}]

school_old_name_ids = list(school_old_name['school_id'])

len(degree_grant)

# let's find out the number of people each of these schools hosted
# to see how much it matters that I have these schools or not

degrees_granted = []  # will hold pairs: (school_id, # of degrees granted)
for s in list(school['school_id']):
    degrees_granted.append((s, len(degree_grant[degree_grant['school']==s])))
