==> 2-Map visualizations.ipynb <==
# Here we will build map visualizations using Basemap,
# and use FFMPEG at the command line to build animations 
# from these Basemap files.

# All the functions for mapping have been saved in the files 
# mgp_functions.py and mgp_map.py (which imports mgp_functions.py).
from mgp_map import *

# We have loaded degree in importing mgp_map.
# We need to prepare degree_with_year: only those degrees with years on them.
degree_with_year = degree[degree['year']>-1].copy()

# This will build a map over all time.
ten_year_ranges = build_year_ranges(first=1290, last=2019, inc=10, over=5)
binned_degrees  = put_data_under_year_ranges(list(degree_with_year['degree_id']), 
                                             list(degree_with_year['year']), 
                                             ten_year_ranges)    
binned_schools = bin_schools_by_time_frame(binned_degrees)
binned_schools_map_all = restructure_schools_for_map(binned_schools)

# total number of errors: 4402 out of 444151 placed: 
# this means that 1% of the mapped schools failed to map (didn't have coordinates).

generate_mgp_map(school_freq_dict=binned_schools_map_all,
                 folder="mgp_img/", fileprefix="all_mgp_year", 
                 title_prefix="All MGP dissertations: ", max_size=100)

# run the following in the terminal 
# to generate the animation via ffmpeg in bash
!./make_anim 3 all_mgp_year.mp4

# This will build a map of aggregated dissertations (so the dots persist) over all time.
binned_schools_map_all_agg = aggs(degree_with_year, f=1290, l=2019, i=10, o=5)

generate_mgp_map(school_freq_dict=binned_schools_map_all_agg,
                 folder="mgp_img/", fileprefix="all_mgp_year_agg", 
                 title_prefix="All MGP dissertations (aggregate): ", max_size=5000)

# run the following in the terminal 
# to generate the animation via ffmpeg in bash
!./make_anim 3 all_mgp_year_agg.mp4
